requirements.txt
pyperclip
pyperclipimg
pillow
websockets